backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
groq==1.0.0
python-dotenv==1.0.0
